+++ client/grpc/cmtservice/service.go
@@ -97,12 +97,7 @@ func (s queryServer) GetBlockByHeight(ctx context.Context, req *GetBlockByHeight
 		return nil, status.Error(codes.InvalidArgument, "requested block height is bigger then the chain length")
 	}
 
-	protoBlockID, protoBlock, err := GetProtoBlock(ctx, s.rpc, &req.Height)
-	if err != nil {
-		return nil, err
-	}
-
-	sdkBlock, err := convertBlock(protoBlock, s.consensusCodec)
+	protoBlockID, protoBlock, err := GetProtoBlock(ctx, s.clientCtx, &req.Height)
 	if err != nil {
 		return nil, err
 	}