+++ client/grpc/tmservice/block.go
@@ -1,0 +1,34 @@
+package tmservice
+
+import (
+	"context"
+
+	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
+	ctypes "github.com/tendermint/tendermint/rpc/core/types"
+
+	"github.com/cosmos/cosmos-sdk/client"
+)
+
+func getBlock(ctx context.Context, clientCtx client.Context, height *int64) (*ctypes.ResultBlock, error) {
+	// get the node
+	node, err := clientCtx.GetNode()
+	if err != nil {
+		return nil, err
+	}
+
+	return node.Block(ctx, height)
+}
+
+func GetProtoBlock(ctx context.Context, clientCtx client.Context, height *int64) (tmproto.BlockID, *tmproto.Block, error) {
+	block, err := getBlock(ctx, clientCtx, height)
+	if err != nil {
+		return tmproto.BlockID{}, nil, err
+	}
+	protoBlock, err := block.Block.ToProto()
+	if err != nil {
+		return tmproto.BlockID{}, nil, err
+	}
+	protoBlockId := block.BlockID.ToProto()
+
+	return protoBlockId, protoBlock, nil
+}