+++ types/tx/service.pb.go
@@ -10,8 +10,11 @@ import (
 	_ "github.com/cosmos/cosmos-proto"
 	types "github.com/cosmos/cosmos-sdk/types"
 	query "github.com/cosmos/cosmos-sdk/types/query"
-	grpc1 "github.com/cosmos/gogoproto/grpc"
-	proto "github.com/cosmos/gogoproto/proto"
+	_ "github.com/gogo/protobuf/gogoproto"
+	grpc1 "github.com/gogo/protobuf/grpc"
+	proto "github.com/gogo/protobuf/proto"
+	golang_proto "github.com/golang/protobuf/proto"
+	types1 "github.com/tendermint/tendermint/proto/tendermint/types"
 	_ "google.golang.org/genproto/googleapis/api/annotations"
 	grpc "google.golang.org/grpc"
 	codes "google.golang.org/grpc/codes"
@@ -109,12 +112,9 @@ func (BroadcastMode) EnumDescriptor() ([]byte, []int) {
 // RPC method.
 type GetTxsEventRequest struct {
 	// events is the list of transaction event type.
-	// Deprecated: post v0.47.x use query instead, which should contain a valid
-	// events query.
-	Events []string `protobuf:"bytes,1,rep,name=events,proto3" json:"events,omitempty"` // Deprecated: Do not use.
+	Events []string `protobuf:"bytes,1,rep,name=events,proto3" json:"events,omitempty"`
 	// pagination defines a pagination for the request.
-	// Deprecated: post v0.46.x use page and limit instead.
-	Pagination *query.PageRequest `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"` // Deprecated: Do not use.
+	Pagination *query.PageRequest `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
 	OrderBy    OrderBy            `protobuf:"varint,3,opt,name=order_by,json=orderBy,proto3,enum=cosmos.tx.v1beta1.OrderBy" json:"order_by,omitempty"`
 	// page is the page number to query, starts at 1. If not provided, will
 	// default to first page.
@@ -212,10 +212,7 @@ type GetTxsEventResponse struct {
 	// tx_responses is the list of queried TxResponses.
 	TxResponses []*types.TxResponse `protobuf:"bytes,2,rep,name=tx_responses,json=txResponses,proto3" json:"tx_responses,omitempty"`
 	// pagination defines a pagination for the response.
-	// Deprecated: post v0.46.x use total instead.
-	Pagination *query.PageResponse `protobuf:"bytes,3,opt,name=pagination,proto3" json:"pagination,omitempty"` // Deprecated: Do not use.
-	// total is total number of results available
-	Total uint64 `protobuf:"varint,4,opt,name=total,proto3" json:"total,omitempty"`
+	Pagination *query.PageResponse `protobuf:"bytes,3,opt,name=pagination,proto3" json:"pagination,omitempty"`
 }
 
 func (m *GetTxsEventResponse) Reset()         { *m = GetTxsEventResponse{} }
@@ -602,6 +599,8 @@ func (m *GetTxResponse) GetTxResponse() *types.TxResponse {
 
 // GetBlockWithTxsRequest is the request type for the Service.GetBlockWithTxs
 // RPC method.
+//
+// Since: cosmos-sdk 0.45.2
 type GetBlockWithTxsRequest struct {
 	// height is the height of the block to query.
 	Height int64 `protobuf:"varint,1,opt,name=height,proto3" json:"height,omitempty"`
@@ -647,6 +646,205 @@ func (m *GetBlockWithTxsRequest) GetHeight() int64 {
 		return m.Height
 	}
 	return 0
+}
+
+func (m *GetBlockWithTxsRequest) GetPagination() *query.PageRequest {
+	if m != nil {
+		return m.Pagination
+	}
+	return nil
+}
+
+// GetBlockWithTxsResponse is the response type for the Service.GetBlockWithTxs method.
+//
+// Since: cosmos-sdk 0.45.2
+type GetBlockWithTxsResponse struct {
+	// txs are the transactions in the block.
+	Txs     []*Tx           `protobuf:"bytes,1,rep,name=txs,proto3" json:"txs,omitempty"`
+	BlockId *types1.BlockID `protobuf:"bytes,2,opt,name=block_id,json=blockId,proto3" json:"block_id,omitempty"`
+	Block   *types1.Block   `protobuf:"bytes,3,opt,name=block,proto3" json:"block,omitempty"`
+	// pagination defines a pagination for the response.
+	Pagination *query.PageResponse `protobuf:"bytes,4,opt,name=pagination,proto3" json:"pagination,omitempty"`
+}
+
+func (m *GetBlockWithTxsResponse) Reset()         { *m = GetBlockWithTxsResponse{} }
+func (m *GetBlockWithTxsResponse) String() string { return proto.CompactTextString(m) }
+func (*GetBlockWithTxsResponse) ProtoMessage()    {}
+func (*GetBlockWithTxsResponse) Descriptor() ([]byte, []int) {
+	return fileDescriptor_e0b00a618705eca7, []int{9}
+}
+func (m *GetBlockWithTxsResponse) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *GetBlockWithTxsResponse) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_GetBlockWithTxsResponse.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *GetBlockWithTxsResponse) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_GetBlockWithTxsResponse.Merge(m, src)
+}
+func (m *GetBlockWithTxsResponse) XXX_Size() int {
+	return m.Size()
+}
+func (m *GetBlockWithTxsResponse) XXX_DiscardUnknown() {
+	xxx_messageInfo_GetBlockWithTxsResponse.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_GetBlockWithTxsResponse proto.InternalMessageInfo
+
+func (m *GetBlockWithTxsResponse) GetTxs() []*Tx {
+	if m != nil {
+		return m.Txs
+	}
+	return nil
+}
+
+func (m *GetBlockWithTxsResponse) GetBlockId() *types1.BlockID {
+	if m != nil {
+		return m.BlockId
+	}
+	return nil
+}
+
+func (m *GetBlockWithTxsResponse) GetBlock() *types1.Block {
+	if m != nil {
+		return m.Block
+	}
+	return nil
+}
+
+func (m *GetBlockWithTxsResponse) GetPagination() *query.PageResponse {
+	if m != nil {
+		return m.Pagination
+	}
+	return nil
+}
+
+func init() {
+	proto.RegisterEnum("cosmos.tx.v1beta1.OrderBy", OrderBy_name, OrderBy_value)
+	golang_proto.RegisterEnum("cosmos.tx.v1beta1.OrderBy", OrderBy_name, OrderBy_value)
+	proto.RegisterEnum("cosmos.tx.v1beta1.BroadcastMode", BroadcastMode_name, BroadcastMode_value)
+	golang_proto.RegisterEnum("cosmos.tx.v1beta1.BroadcastMode", BroadcastMode_name, BroadcastMode_value)
+	proto.RegisterType((*GetTxsEventRequest)(nil), "cosmos.tx.v1beta1.GetTxsEventRequest")
+	golang_proto.RegisterType((*GetTxsEventRequest)(nil), "cosmos.tx.v1beta1.GetTxsEventRequest")
+	proto.RegisterType((*GetTxsEventResponse)(nil), "cosmos.tx.v1beta1.GetTxsEventResponse")
+	golang_proto.RegisterType((*GetTxsEventResponse)(nil), "cosmos.tx.v1beta1.GetTxsEventResponse")
+	proto.RegisterType((*BroadcastTxRequest)(nil), "cosmos.tx.v1beta1.BroadcastTxRequest")
+	golang_proto.RegisterType((*BroadcastTxRequest)(nil), "cosmos.tx.v1beta1.BroadcastTxRequest")
+	proto.RegisterType((*BroadcastTxResponse)(nil), "cosmos.tx.v1beta1.BroadcastTxResponse")
+	golang_proto.RegisterType((*BroadcastTxResponse)(nil), "cosmos.tx.v1beta1.BroadcastTxResponse")
+	proto.RegisterType((*SimulateRequest)(nil), "cosmos.tx.v1beta1.SimulateRequest")
+	golang_proto.RegisterType((*SimulateRequest)(nil), "cosmos.tx.v1beta1.SimulateRequest")
+	proto.RegisterType((*SimulateResponse)(nil), "cosmos.tx.v1beta1.SimulateResponse")
+	golang_proto.RegisterType((*SimulateResponse)(nil), "cosmos.tx.v1beta1.SimulateResponse")
+	proto.RegisterType((*GetTxRequest)(nil), "cosmos.tx.v1beta1.GetTxRequest")
+	golang_proto.RegisterType((*GetTxRequest)(nil), "cosmos.tx.v1beta1.GetTxRequest")
+	proto.RegisterType((*GetTxResponse)(nil), "cosmos.tx.v1beta1.GetTxResponse")
+	golang_proto.RegisterType((*GetTxResponse)(nil), "cosmos.tx.v1beta1.GetTxResponse")
+	proto.RegisterType((*GetBlockWithTxsRequest)(nil), "cosmos.tx.v1beta1.GetBlockWithTxsRequest")
+	golang_proto.RegisterType((*GetBlockWithTxsRequest)(nil), "cosmos.tx.v1beta1.GetBlockWithTxsRequest")
+	proto.RegisterType((*GetBlockWithTxsResponse)(nil), "cosmos.tx.v1beta1.GetBlockWithTxsResponse")
+	golang_proto.RegisterType((*GetBlockWithTxsResponse)(nil), "cosmos.tx.v1beta1.GetBlockWithTxsResponse")
+}
+
+func (m *GetBlockWithTxsRequest) Reset()         { *m = GetBlockWithTxsRequest{} }
+func (m *GetBlockWithTxsRequest) String() string { return proto.CompactTextString(m) }
+func (*GetBlockWithTxsRequest) ProtoMessage()    {}
+func (*GetBlockWithTxsRequest) Descriptor() ([]byte, []int) {
+	return fileDescriptor_e0b00a618705eca7, []int{8}
+}
+
+var fileDescriptor_e0b00a618705eca7 = []byte{
+	// 976 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xac, 0x56, 0x4f, 0x6f, 0x1a, 0x47,
+	0x14, 0xf7, 0x2e, 0xb6, 0x21, 0x0f, 0x3b, 0x21, 0x63, 0xd7, 0x26, 0x24, 0xc5, 0x64, 0x53, 0x6c,
+	0x07, 0xc9, 0xbb, 0x0a, 0x4d, 0xa5, 0xaa, 0xea, 0xc5, 0xfc, 0x89, 0x8b, 0xda, 0x84, 0x68, 0x70,
+	0x15, 0xa5, 0xaa, 0x84, 0x16, 0x98, 0x2c, 0xab, 0x98, 0x1d, 0xbc, 0x33, 0x58, 0x8b, 0x1c, 0xab,
+	0x52, 0x8f, 0x3d, 0x55, 0xed, 0xa1, 0x1f, 0xa2, 0x5f, 0xa2, 0xc7, 0x1e, 0x2d, 0xf5, 0xd2, 0x63,
+	0x65, 0xf7, 0x03, 0xf4, 0x23, 0x54, 0x3b, 0x3b, 0xc0, 0x82, 0x97, 0x38, 0x8d, 0x7a, 0x81, 0x99,
+	0x9d, 0xdf, 0x7b, 0xef, 0x37, 0xbf, 0xd9, 0xdf, 0x9b, 0x85, 0xad, 0x36, 0x65, 0x3d, 0xca, 0x0c,
+	0xee, 0x19, 0x27, 0x8f, 0x5a, 0x84, 0x9b, 0x8f, 0x0c, 0x46, 0xdc, 0x13, 0xbb, 0x4d, 0xf4, 0xbe,
+	0x4b, 0x39, 0x45, 0xb7, 0x03, 0x80, 0xce, 0x3d, 0x5d, 0x02, 0x32, 0xf7, 0x2c, 0x4a, 0xad, 0x23,
+	0x62, 0x98, 0x7d, 0xdb, 0x30, 0x1d, 0x87, 0x72, 0x93, 0xdb, 0xd4, 0x61, 0x41, 0x40, 0xe6, 0x81,
+	0xcc, 0xd8, 0x32, 0x19, 0x31, 0xcc, 0x56, 0xdb, 0x1e, 0x27, 0xf6, 0x27, 0x12, 0x94, 0xb9, 0x5a,
+	0x96, 0x7b, 0x72, 0x6d, 0xdd, 0xa2, 0x16, 0x15, 0x43, 0xc3, 0x1f, 0xc9, 0xa7, 0x85, 0x70, 0xda,
+	0xe3, 0x01, 0x71, 0x87, 0xe3, 0xc8, 0xbe, 0x69, 0xd9, 0x8e, 0xe0, 0x20, 0xb1, 0xf7, 0x38, 0x71,
+	0x3a, 0xc4, 0xed, 0xd9, 0x0e, 0x37, 0xf8, 0xb0, 0x4f, 0x98, 0xd1, 0x3a, 0xa2, 0xed, 0xd7, 0x73,
+	0x57, 0xc5, 0x6f, 0xb0, 0xaa, 0xfd, 0xaa, 0x00, 0x3a, 0x20, 0xfc, 0xd0, 0x63, 0xd5, 0x13, 0xe2,
+	0x70, 0x4c, 0x8e, 0x07, 0x84, 0x71, 0xb4, 0x01, 0xcb, 0xc4, 0x9f, 0xb3, 0xb4, 0x92, 0x8b, 0xed,
+	0xde, 0xc0, 0x72, 0x86, 0x9e, 0x00, 0x4c, 0xca, 0xa7, 0xd5, 0x9c, 0xb2, 0x9b, 0x2c, 0x6e, 0xeb,
+	0x52, 0x33, 0x9f, 0xab, 0x2e, 0xb8, 0x8e, 0xb4, 0xd3, 0x9f, 0x9b, 0x16, 0x91, 0x39, 0x71, 0x28,
+	0x12, 0x7d, 0x02, 0x09, 0xea, 0x76, 0x88, 0xdb, 0x6c, 0x0d, 0xd3, 0xb1, 0x9c, 0xb2, 0x7b, 0xb3,
+	0x98, 0xd1, 0xaf, 0x28, 0xaf, 0xd7, 0x7d, 0x48, 0x69, 0x88, 0xe3, 0x34, 0x18, 0x68, 0xe7, 0x0a,
+	0xac, 0x4d, 0xb1, 0x65, 0x7d, 0xea, 0x30, 0x82, 0x76, 0x20, 0xc6, 0xbd, 0x80, 0x6b, 0xb2, 0xf8,
+	0x41, 0x44, 0xa6, 0x43, 0x0f, 0xfb, 0x08, 0x74, 0x00, 0x2b, 0xdc, 0x6b, 0xba, 0x32, 0x8e, 0xa5,
+	0x55, 0x11, 0xf1, 0xd1, 0xd4, 0x0e, 0xc4, 0xb9, 0x85, 0x02, 0x25, 0x18, 0x27, 0xf9, 0x78, 0xec,
+	0x27, 0x0a, 0x0b, 0x11, 0x13, 0x42, 0xec, 0x5c, 0x2b, 0x84, 0xcc, 0x14, 0x0a, 0xd5, 0x08, 0xa0,
+	0x92, 0x4b, 0xcd, 0x4e, 0xdb, 0x64, 0xdc, 0x2f, 0x16, 0xe8, 0x7f, 0x07, 0x12, 0xdc, 0x6b, 0xb6,
+	0x86, 0x9c, 0xf8, 0xbb, 0x52, 0x76, 0x57, 0x70, 0x9c, 0x7b, 0x25, 0x7f, 0x8a, 0x1e, 0xc3, 0x62,
+	0x8f, 0x76, 0x88, 0x10, 0xff, 0x66, 0x31, 0x17, 0xb1, 0xd9, 0x71, 0xbe, 0xa7, 0xb4, 0x43, 0xb0,
+	0x40, 0x6b, 0xdf, 0xc2, 0xda, 0x54, 0x19, 0x29, 0x5c, 0x15, 0x92, 0x21, 0x3d, 0x44, 0xa9, 0x77,
+	0x95, 0x03, 0x26, 0x72, 0x68, 0x2f, 0xe0, 0x56, 0xc3, 0xee, 0x0d, 0x8e, 0x4c, 0x3e, 0x3a, 0x6d,
+	0xf4, 0x10, 0x54, 0xee, 0xc9, 0x84, 0xd1, 0x27, 0x52, 0x52, 0xd3, 0x0a, 0x56, 0xb9, 0x37, 0xb5,
+	0x59, 0x75, 0x6a, 0xb3, 0xda, 0x0f, 0x0a, 0xa4, 0x26, 0x99, 0x25, 0xe9, 0xcf, 0x21, 0x61, 0x99,
+	0xac, 0x69, 0x3b, 0xaf, 0xa8, 0x2c, 0x70, 0x7f, 0x3e, 0xe3, 0x03, 0x93, 0xd5, 0x9c, 0x57, 0x14,
+	0xc7, 0xad, 0x60, 0x80, 0x3e, 0x85, 0x65, 0x97, 0xb0, 0xc1, 0x11, 0x97, 0xaf, 0x6f, 0x6e, 0x7e,
+	0x2c, 0x16, 0x38, 0x2c, 0xf1, 0x9a, 0x06, 0x2b, 0xe2, 0xe5, 0x1b, 0x6d, 0x11, 0xc1, 0x62, 0xd7,
+	0x64, 0x5d, 0xc1, 0xe1, 0x06, 0x16, 0x63, 0xed, 0x0c, 0x56, 0x25, 0x46, 0x92, 0xcd, 0x5f, 0xab,
+	0x83, 0xd0, 0x60, 0xe6, 0x20, 0xd4, 0xf7, 0x3c, 0x08, 0x0f, 0x36, 0x0e, 0x08, 0x2f, 0xf9, 0xf6,
+	0x7f, 0x61, 0xf3, 0xee, 0xa1, 0xc7, 0x42, 0x8e, 0xee, 0x12, 0xdb, 0xea, 0x72, 0xc1, 0x25, 0x86,
+	0xe5, 0xec, 0xff, 0x72, 0xb4, 0xf6, 0x8f, 0x02, 0x9b, 0x57, 0x4a, 0xff, 0x57, 0x7b, 0x3e, 0x86,
+	0x84, 0x68, 0x5d, 0x4d, 0xbb, 0x23, 0xa9, 0xdc, 0xd1, 0x27, 0xed, 0x4b, 0x0f, 0x1a, 0x97, 0x28,
+	0x51, 0xab, 0xe0, 0xb8, 0x80, 0xd6, 0x3a, 0x68, 0x0f, 0x96, 0xc4, 0x50, 0xda, 0x70, 0x73, 0x4e,
+	0x08, 0x0e, 0x50, 0x33, 0xd6, 0x5d, 0x7c, 0x6f, 0xeb, 0x16, 0xbe, 0x80, 0xb8, 0xec, 0x50, 0x28,
+	0x0d, 0xeb, 0x75, 0x5c, 0xa9, 0xe2, 0x66, 0xe9, 0x65, 0xf3, 0xeb, 0x67, 0x8d, 0xe7, 0xd5, 0x72,
+	0xed, 0x49, 0xad, 0x5a, 0x49, 0x2d, 0xa0, 0x14, 0xac, 0x8c, 0x57, 0xf6, 0x1b, 0xe5, 0x94, 0x82,
+	0x6e, 0xc3, 0xea, 0xf8, 0x49, 0xa5, 0xda, 0x28, 0xa7, 0xd4, 0xc2, 0x1b, 0x58, 0x9d, 0x32, 0x2d,
+	0xca, 0x42, 0xa6, 0x84, 0xeb, 0xfb, 0x95, 0xf2, 0x7e, 0xe3, 0xb0, 0xf9, 0xb4, 0x5e, 0xa9, 0xce,
+	0x64, 0x4d, 0xc3, 0xfa, 0xcc, 0x7a, 0xe9, 0xab, 0x7a, 0xf9, 0xcb, 0x94, 0x82, 0x36, 0x61, 0x6d,
+	0x66, 0xa5, 0xf1, 0xf2, 0x59, 0x39, 0xa5, 0x46, 0x84, 0xec, 0x8b, 0x95, 0x58, 0xf1, 0xa7, 0x25,
+	0x88, 0x37, 0x82, 0x5b, 0x10, 0x9d, 0x42, 0x62, 0xe4, 0x37, 0xa4, 0x45, 0x9c, 0xd4, 0x8c, 0xcd,
+	0x33, 0x0f, 0xde, 0x8a, 0x91, 0x6f, 0xe5, 0xf6, 0xf7, 0x7f, 0xfc, 0xfd, 0xb3, 0x9a, 0xd3, 0xee,
+	0x1a, 0x11, 0xd7, 0xaf, 0x04, 0x7f, 0xa6, 0x14, 0xd0, 0x31, 0x2c, 0x09, 0xf3, 0xa0, 0xad, 0x88,
+	0xac, 0x61, 0xeb, 0x65, 0x72, 0xf3, 0x01, 0xb2, 0x66, 0x5e, 0xd4, 0xdc, 0x42, 0x1f, 0x1a, 0x51,
+	0x77, 0x2f, 0x33, 0x4e, 0x7d, 0xbb, 0x9e, 0xa1, 0xef, 0x20, 0x19, 0xea, 0x8b, 0x28, 0xff, 0xb6,
+	0x76, 0x3a, 0x29, 0xbf, 0x7d, 0x1d, 0x4c, 0x92, 0xb8, 0x2f, 0x48, 0xdc, 0xd5, 0x36, 0xa2, 0x49,
+	0xf8, 0x7b, 0x7e, 0x03, 0xc9, 0xd0, 0x8d, 0x16, 0x49, 0xe0, 0xea, 0xfd, 0x1c, 0x49, 0x20, 0xe2,
+	0x62, 0xd4, 0xb2, 0x82, 0x40, 0x1a, 0xcd, 0x21, 0x80, 0x7e, 0x51, 0xe0, 0xd6, 0x8c, 0x6b, 0xd1,
+	0xc3, 0xe8, 0xdc, 0x11, 0x4d, 0x25, 0x53, 0x78, 0x17, 0xa8, 0xa4, 0xb2, 0x27, 0xa8, 0xec, 0xa0,
+	0xfc, 0x9c, 0x03, 0x11, 0xe6, 0x34, 0x4e, 0x83, 0xb6, 0x74, 0x56, 0x2a, 0xff, 0x7e, 0x91, 0x55,
+	0xce, 0x2f, 0xb2, 0xca, 0x5f, 0x17, 0x59, 0xe5, 0xc7, 0xcb, 0xec, 0xc2, 0x6f, 0x97, 0x59, 0xe5,
+	0xfc, 0x32, 0xbb, 0xf0, 0xe7, 0x65, 0x76, 0xe1, 0x9b, 0xbc, 0x65, 0xf3, 0xee, 0xa0, 0xa5, 0xb7,
+	0x69, 0x6f, 0x94, 0x2e, 0xf8, 0xdb, 0x63, 0x9d, 0xd7, 0xa3, 0xcf, 0x1c, 0xaf, 0xb5, 0x2c, 0x3e,
+	0x72, 0x3e, 0xfe, 0x37, 0x00, 0x00, 0xff, 0xff, 0xb5, 0xaa, 0xba, 0x0d, 0xf7, 0x09, 0x00, 0x00,
+}
+
+var xxx_messageInfo_GetBlockWithTxsRequest proto.InternalMessageInfo
+
+// ServiceClient is the client API for Service service.
+//
+// For semantics around ctx use and closing/ending streaming RPCs, please refer to https://godoc.org/google.golang.org/grpc#ClientConn.NewStream.
+type ServiceClient interface {
+	// Simulate simulates executing a transaction for estimating gas usage.
+	Simulate(ctx context.Context, in *SimulateRequest, opts ...grpc.CallOption) (*SimulateResponse, error)
+	// GetTx fetches a tx by hash.
+	GetTx(ctx context.Context, in *GetTxRequest, opts ...grpc.CallOption) (*GetTxResponse, error)
+	// BroadcastTx broadcast transaction.
+	BroadcastTx(ctx context.Context, in *BroadcastTxRequest, opts ...grpc.CallOption) (*BroadcastTxResponse, error)
+	// GetTxsEvent fetches txs by event.
+	GetTxsEvent(ctx context.Context, in *GetTxsEventRequest, opts ...grpc.CallOption) (*GetTxsEventResponse, error)
+	// GetBlockWithTxs fetches a block with decoded txs.
+	//
+	// Since: cosmos-sdk 0.45.2
+	GetBlockWithTxs(ctx context.Context, in *GetBlockWithTxsRequest, opts ...grpc.CallOption) (*GetBlockWithTxsResponse, error)
 }
 
 func (m *GetBlockWithTxsRequest) GetPagination() *query.PageRequest {
@@ -721,11 +919,29 @@ func (m *GetBlockWithTxsResponse) GetBlock() *v1.Block {
 	return nil
 }
 
-func (m *GetBlockWithTxsResponse) GetPagination() *query.PageResponse {
-	if m != nil {
-		return m.Pagination
+func (c *serviceClient) GetBlockWithTxs(ctx context.Context, in *GetBlockWithTxsRequest, opts ...grpc.CallOption) (*GetBlockWithTxsResponse, error) {
+	out := new(GetBlockWithTxsResponse)
+	err := c.cc.Invoke(ctx, "/cosmos.tx.v1beta1.Service/GetBlockWithTxs", in, out, opts...)
+	if err != nil {
+		return nil, err
 	}
-	return nil
+	return out, nil
+}
+
+// ServiceServer is the server API for Service service.
+type ServiceServer interface {
+	// Simulate simulates executing a transaction for estimating gas usage.
+	Simulate(context.Context, *SimulateRequest) (*SimulateResponse, error)
+	// GetTx fetches a tx by hash.
+	GetTx(context.Context, *GetTxRequest) (*GetTxResponse, error)
+	// BroadcastTx broadcast transaction.
+	BroadcastTx(context.Context, *BroadcastTxRequest) (*BroadcastTxResponse, error)
+	// GetTxsEvent fetches txs by event.
+	GetTxsEvent(context.Context, *GetTxsEventRequest) (*GetTxsEventResponse, error)
+	// GetBlockWithTxs fetches a block with decoded txs.
+	//
+	// Since: cosmos-sdk 0.45.2
+	GetBlockWithTxs(context.Context, *GetBlockWithTxsRequest) (*GetBlockWithTxsResponse, error)
 }
 
 // TxDecodeRequest is the request type for the Service.TxDecode
@@ -759,11 +975,12 @@ func (m *TxDecodeRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, err
 func (m *TxDecodeRequest) XXX_Merge(src proto.Message) {
 	xxx_messageInfo_TxDecodeRequest.Merge(m, src)
 }
-func (m *TxDecodeRequest) XXX_Size() int {
-	return m.Size()
+func (*UnimplementedServiceServer) GetBlockWithTxs(ctx context.Context, req *GetBlockWithTxsRequest) (*GetBlockWithTxsResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetBlockWithTxs not implemented")
 }
-func (m *TxDecodeRequest) XXX_DiscardUnknown() {
-	xxx_messageInfo_TxDecodeRequest.DiscardUnknown(m)
+
+func RegisterServiceServer(s grpc1.Server, srv ServiceServer) {
+	s.RegisterService(&_Service_serviceDesc, srv)
 }
 
 var xxx_messageInfo_TxDecodeRequest proto.InternalMessageInfo
@@ -822,11 +1039,51 @@ func (m *TxDecodeResponse) GetTx() *Tx {
 	return nil
 }
 
-// TxEncodeRequest is the request type for the Service.TxEncode
-// RPC method.
-type TxEncodeRequest struct {
-	// tx is the transaction to encode.
-	Tx *Tx `protobuf:"bytes,1,opt,name=tx,proto3" json:"tx,omitempty"`
+func _Service_GetBlockWithTxs_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(GetBlockWithTxsRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(ServiceServer).GetBlockWithTxs(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/cosmos.tx.v1beta1.Service/GetBlockWithTxs",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(ServiceServer).GetBlockWithTxs(ctx, req.(*GetBlockWithTxsRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+var _Service_serviceDesc = grpc.ServiceDesc{
+	ServiceName: "cosmos.tx.v1beta1.Service",
+	HandlerType: (*ServiceServer)(nil),
+	Methods: []grpc.MethodDesc{
+		{
+			MethodName: "Simulate",
+			Handler:    _Service_Simulate_Handler,
+		},
+		{
+			MethodName: "GetTx",
+			Handler:    _Service_GetTx_Handler,
+		},
+		{
+			MethodName: "BroadcastTx",
+			Handler:    _Service_BroadcastTx_Handler,
+		},
+		{
+			MethodName: "GetTxsEvent",
+			Handler:    _Service_GetTxsEvent_Handler,
+		},
+		{
+			MethodName: "GetBlockWithTxs",
+			Handler:    _Service_GetBlockWithTxs_Handler,
+		},
+	},
+	Streams:  []grpc.StreamDesc{},
+	Metadata: "cosmos/tx/v1beta1/service.proto",
 }
 
 func (m *TxEncodeRequest) Reset()         { *m = TxEncodeRequest{} }
@@ -1315,11 +1572,126 @@ func (c *serviceClient) TxEncode(ctx context.Context, in *TxEncodeRequest, opts 
 	return out, nil
 }
 
-func (c *serviceClient) TxEncodeAmino(ctx context.Context, in *TxEncodeAminoRequest, opts ...grpc.CallOption) (*TxEncodeAminoResponse, error) {
-	out := new(TxEncodeAminoResponse)
-	err := c.cc.Invoke(ctx, "/cosmos.tx.v1beta1.Service/TxEncodeAmino", in, out, opts...)
+func (m *GetBlockWithTxsRequest) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
 	if err != nil {
 		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *GetBlockWithTxsRequest) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *GetBlockWithTxsRequest) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if m.Pagination != nil {
+		{
+			size, err := m.Pagination.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintService(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x12
+	}
+	if m.Height != 0 {
+		i = encodeVarintService(dAtA, i, uint64(m.Height))
+		i--
+		dAtA[i] = 0x8
+	}
+	return len(dAtA) - i, nil
+}
+
+func (m *GetBlockWithTxsResponse) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *GetBlockWithTxsResponse) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *GetBlockWithTxsResponse) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if m.Pagination != nil {
+		{
+			size, err := m.Pagination.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintService(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x22
+	}
+	if m.Block != nil {
+		{
+			size, err := m.Block.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintService(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x1a
+	}
+	if m.BlockId != nil {
+		{
+			size, err := m.BlockId.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintService(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x12
+	}
+	if len(m.Txs) > 0 {
+		for iNdEx := len(m.Txs) - 1; iNdEx >= 0; iNdEx-- {
+			{
+				size, err := m.Txs[iNdEx].MarshalToSizedBuffer(dAtA[:i])
+				if err != nil {
+					return 0, err
+				}
+				i -= size
+				i = encodeVarintService(dAtA, i, uint64(size))
+			}
+			i--
+			dAtA[i] = 0xa
+		}
+	}
+	return len(dAtA) - i, nil
+}
+
+func encodeVarintService(dAtA []byte, offset int, v uint64) int {
+	offset -= sovService(v)
+	base := offset
+	for v >= 1<<7 {
+		dAtA[offset] = uint8(v&0x7f | 0x80)
+		v >>= 7
+		offset++
 	}
 	return out, nil
 }
@@ -3458,7 +3830,57 @@ func (m *SimulateResponse) Unmarshal(dAtA []byte) error {
 	if iNdEx > l {
 		return io.ErrUnexpectedEOF
 	}
-	return nil
+	return n
+}
+
+func (m *GetBlockWithTxsRequest) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	if m.Height != 0 {
+		n += 1 + sovService(uint64(m.Height))
+	}
+	if m.Pagination != nil {
+		l = m.Pagination.Size()
+		n += 1 + l + sovService(uint64(l))
+	}
+	return n
+}
+
+func (m *GetBlockWithTxsResponse) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	if len(m.Txs) > 0 {
+		for _, e := range m.Txs {
+			l = e.Size()
+			n += 1 + l + sovService(uint64(l))
+		}
+	}
+	if m.BlockId != nil {
+		l = m.BlockId.Size()
+		n += 1 + l + sovService(uint64(l))
+	}
+	if m.Block != nil {
+		l = m.Block.Size()
+		n += 1 + l + sovService(uint64(l))
+	}
+	if m.Pagination != nil {
+		l = m.Pagination.Size()
+		n += 1 + l + sovService(uint64(l))
+	}
+	return n
+}
+
+func sovService(x uint64) (n int) {
+	return (math_bits.Len64(x|1) + 6) / 7
+}
+func sozService(x uint64) (n int) {
+	return sovService(uint64((x << 1) ^ uint64((int64(x) >> 63))))
 }
 func (m *GetTxRequest) Unmarshal(dAtA []byte) error {
 	l := len(dAtA)
@@ -4611,6 +5033,303 @@ func (m *TxDecodeAminoResponse) Unmarshal(dAtA []byte) error {
 				return io.ErrUnexpectedEOF
 			}
 			m.AminoJson = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		default:
+			iNdEx = preIndex
+			skippy, err := skipService(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthService
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *GetBlockWithTxsRequest) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowService
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: GetBlockWithTxsRequest: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: GetBlockWithTxsRequest: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Height", wireType)
+			}
+			m.Height = 0
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowService
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				m.Height |= int64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Pagination", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowService
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthService
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthService
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Pagination == nil {
+				m.Pagination = &query.PageRequest{}
+			}
+			if err := m.Pagination.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		default:
+			iNdEx = preIndex
+			skippy, err := skipService(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthService
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *GetBlockWithTxsResponse) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowService
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: GetBlockWithTxsResponse: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: GetBlockWithTxsResponse: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Txs", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowService
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthService
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthService
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.Txs = append(m.Txs, &Tx{})
+			if err := m.Txs[len(m.Txs)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field BlockId", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowService
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthService
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthService
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.BlockId == nil {
+				m.BlockId = &types1.BlockID{}
+			}
+			if err := m.BlockId.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 3:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Block", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowService
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthService
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthService
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Block == nil {
+				m.Block = &types1.Block{}
+			}
+			if err := m.Block.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 4:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Pagination", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowService
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthService
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthService
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Pagination == nil {
+				m.Pagination = &query.PageResponse{}
+			}
+			if err := m.Pagination.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
 			iNdEx = postIndex
 		default:
 			iNdEx = preIndex